pagweb/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
boto3==1.16.21
botocore==1.19.21
certifi==2020.6.20
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-heroku==0.3.1
django-storages==1.10.1
filelock==3.0.12
gunicorn==20.0.4
jmespath==0.10.0
lpthw.web==1.1
nose==1.3.7
olefile==0.46
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2
virtualenv==20.0.28
whitenoise==5.2.0
wincertstore==0.2
